test(all): move configs to config folder, and change active profiles test name
