+++ infrastructure/src/test/java/com/kaua/ecommerce/users/CacheGatewayTest.java
@@ -1,5 +1,9 @@
 package com.kaua.ecommerce.users;
 
+import com.kaua.ecommerce.users.config.AmqpTestConfiguration;
+import com.kaua.ecommerce.users.config.CacheCleanUpExtension;
+import com.kaua.ecommerce.users.config.CacheTestConfiguration;
+import com.kaua.ecommerce.users.config.JpaCleanUpExtension;
 import com.kaua.ecommerce.users.infrastructure.Main;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -11,8 +15,8 @@ import java.lang.annotation.*;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
-@ActiveProfiles("test")
+@ActiveProfiles("test-cache-integration")
 @SpringBootTest(classes = { Main.class, AmqpTestConfiguration.class, CacheTestConfiguration.class })
-@ExtendWith(CleanUpExtension.class)
+@ExtendWith({ CacheCleanUpExtension.class, JpaCleanUpExtension.class})
 public @interface CacheGatewayTest {
 }
