+++ infrastructure/src/test/java/com/kaua/ecommerce/users/config/CacheCleanUpExtension.java
@@ -1,0 +1,26 @@
+package com.kaua.ecommerce.users.config;
+
+import com.kaua.ecommerce.users.infrastructure.accounts.persistence.AccountCacheRepository;
+import org.junit.jupiter.api.extension.BeforeEachCallback;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import java.util.Collection;
+import java.util.List;
+
+public class CacheCleanUpExtension implements BeforeEachCallback {
+
+    @Override
+    public void beforeEach(ExtensionContext context) throws Exception {
+        final var appContext = SpringExtension.getApplicationContext(context);
+
+        cleanUp(List.of(
+                appContext.getBean(AccountCacheRepository.class)
+        ));
+    }
+
+    private void cleanUp(final Collection<CrudRepository> repositories) {
+        repositories.forEach(CrudRepository::deleteAll);
+    }
+}
