+++ infrastructure/src/test/java/com/kaua/ecommerce/users/config/JpaCleanUpExtension.java
@@ -1,4 +1,4 @@
-package com.kaua.ecommerce.users;
+package com.kaua.ecommerce.users.config;
 
 import com.kaua.ecommerce.users.infrastructure.accounts.mail.persistence.AccountMailJpaRepository;
 import com.kaua.ecommerce.users.infrastructure.accounts.persistence.AccountJpaRepository;
